reorg the view base class and default view
There are no behavior changes here - just reorgs and some test cleanups
and additions.  This is part of cleaning things up and making them
more intentionally extendable in prep for working on Parassert.

First, I want to get rid of the `view` subfolder.  This moves the
default view up to a top-level object and moves the base logic
out of its separate file and into the top-level view file.  The
rest of the changes just support these moves and make the test more
up to our latest conventions relative to these changes.
